+++ pipeline/talend_test/ansible/roles/process_files/library/allowed_dir_test.json
@@ -1,0 +1,6 @@
+{
+  "ANSIBLE_MODULE_ARGS": {
+      "allowed_dirs": ["/test"],
+      "test": "/test/some.file"
+    }
+}
