+++ pipeline/talend_test/ansible/roles/process_files/tasks/harvest_results.yaml
@@ -1,0 +1,61 @@
+
+### process action sub-playbook. adds input files and waits for processing to complete
+
+- name: create temporary local directory
+  delegate_to: localhost
+  tempfile:
+    state: directory
+    suffix: build
+  register: new_tmp_dir
+
+
+- name: set tmp directory fact
+  set_fact:
+    tmp_dir: "{{ new_tmp_dir.path}}"
+
+
+- name: debug tmp_dir path
+  debug:
+    msg: "{{ tmp_dir }}"
+
+
+- name: create dirs for pgsql output
+  delegate_to: localhost
+  file: path={{ tmp_dir }}/{{ item.name }} state=directory mode="a+w"
+  with_items: "{{ database_schemas }}"
+
+
+- set_fact:
+    database_user: "{{ db_user | default(ansible_ssh_user) }}"
+
+
+- name: sql select queries
+  delegate_to: localhost
+  psql_harvest_queries:
+    schema_obj: "{{ item }}"
+    tmp_dir: "{{ tmp_dir }}/{{ item.name }}"
+    db_user: "{{ database_user }}"
+    db_host: "{{ hosts }}"
+  loop: "{{ database_schemas }}"
+
+
+- name: remove result dir
+  delegate_to: localhost
+  file:
+    state: absent
+    path: "../harvest_results/{{ hosts }}/{{ name }}/{{ label }}/"
+
+
+- name: create result dir
+  delegate_to: localhost
+  file:
+    path: "../harvest_results/{{ hosts }}/{{ name }}/{{ label }}/"
+    state: directory
+
+
+- name: copy results from temp to result dir
+  delegate_to: localhost
+  copy:
+    src: "{{ tmp_dir }}/{{ item.name }}"
+    dest: "../harvest_results/{{ hosts }}/{{ name }}/{{ label }}/"
+  loop: "{{ database_schemas }}"
