+++ pipeline/talend_test/ansible/roles/process_files/tasks/schema_restore.yaml
@@ -1,0 +1,60 @@
+- name: create schema download directory
+  delegate_to: localhost
+  file:
+    path: /tmp/talend_test_schemas
+    state: directory
+
+
+- name: download schema backup to local
+  delegate_to: localhost
+  aws_s3:
+    bucket: imos-backups
+    object: "{{ schema_restore.object.path }}/{{ item.name }}.dump"
+    dest: "/tmp/talend_test_schemas/{{ item.name }}.dump"
+    mode: get
+  loop: "{{ database_schemas }}"
+
+
+- name: create remote schema upload directory
+  become: yes
+  become_user: root
+  file:
+    path: /mnt/talend_test_schemas
+    state: directory
+    mode: a+rw
+
+
+- name: upload schema backup to remote
+  synchronize:
+    src: "/tmp/talend_test_schemas/{{ item.name }}.dump"
+    dest: /mnt/talend_test_schemas
+  loop: "{{ database_schemas }}"
+
+
+- name: drop schema
+  command: "sudo -u postgres psql -d harvest -c 'DROP SCHEMA IF EXISTS {{ item.name }} CASCADE;'"
+  loop: "{{ database_schemas }}"
+
+
+- name: run pg_restore on schema backup
+  command: "sudo -u postgres pg_restore -d harvest /mnt/talend_test_schemas/{{ item.name }}.dump"
+  loop: "{{ database_schemas }}"
+
+
+- name: remove schema dump from remote disk
+  file:
+    state: absent
+    path: "/mnt/talend_test_schemas/{{ item.name }}.dump"
+  loop: "{{ database_schemas }}"
+
+
+- name: remove schema dump from local disk
+  delegate_to: localhost
+  file:
+    state: absent
+    path: "/tmp/talend_test_schemas/{{ item.name }}.dump"
+  loop: "{{ database_schemas }}"
+
+
+
+
