--- pipeline/talend_test/ansible/roles/talend_test/library/allowed_dir_test.json
@@ -1,6 +1,0 @@
-{
-  "ANSIBLE_MODULE_ARGS": {
-      "allowed_dirs": ["/test"],
-      "test": "/test/some.file"
-    }
-}
