--- pipeline/talend_test/global_vars.yaml
@@ -1,20 +1,0 @@
-
-pipeline_1_process_log: /mnt/ebs/log/data-services/process.log
-
-allowed_log_dirs:
-  - /mnt/ebs/log/data-services
-  - /mnt/ebs/log/pipeline/process
-  - /usr/local/talend/jobs/aodn_wave_nrt-aodn_wave_nrt/log
-  - /usr/local/talend/jobs/soop_auscpr-soop_auscpr/log
-  - /usr/local/talend/jobs/aodn_wamsi_sediment-aodn_wamsi_sediment/log
-  - /usr/local/talend/jobs/aodn_imas_fluorometry-aodn_imas_fluorometry/log
-  - /usr/local/talend/jobs/aodn_bom_sea_level-aodn_bom_sea_level/log
-  - /usr/local/talend/jobs/aatams_acoustic_reporting-aatams_acoustic_reporting/log
-  - /usr/local/talend/jobs/anmn_nrs_bgc-anmn_nrs_bgc/log
-  - /usr/local/talend/jobs/parameters_mapping-parameters_mapping/log
-
-default_timeout: 800
-
-host_buckets:
-  6-nec-hob.emii.org.au: imos-data-pipeline-2
-  9-nec-hob.emii.org.au: imos-data-pipeline-talend7
