+++ pipeline/talend_test/requirements.txt
@@ -1,2 +1,3 @@
-ansible
+ansible>=2.8
 boto3
+psycopg2
