+++ pipeline/talend_test/test_configs/soop_trv/expect/action_0_ADD/soop_trv/soop_trv_trajectory_map.csv
@@ -1,0 +1,2 @@
+trip_id,metadata_uuid,time_coverage_start,time_coverage_end,TURB_b,CPHL_b,TEMP_b,PSAL_b,geom,platform_code,vessel_name
+6997,852FAC8B-393B-543B-E054-0010E056919A,2019-03-28 19:15:56,2019-04-07 21:57:25,f,f,f,t,0102000020E6100000E90000007FBC57AD4C405E402237C30DF8FC2FC0CF6BEC12D5435E40CCD1E3F7369D2FC039D1AE42CA3F5E407B14AE47E1FA2FC04E9CDCEF50405E40BD5296218EF52FC049D74CBED93E5E409B559FABADF82FC0E4F736FDD93F5E4004FF5BC98EED2FC0C2DD59BBED3E5E401956F146E6F12FC08E9257E718405E40F67F0EF3E5E52FC01E5036E50A3F5E402CD49AE61DE72FC0774A07EBFF405E400AA2EE0390FA2FC051888043A83E5E404F401361C3F32FC0DC80CF0F23405E405CC98E8D40DC2FC077BE9F1A2F415E40642310AFEBF72FC02DB29DEFA73E5E406E179AEB34F22FC0CEC29E76F83F5E4048BF7D1D38E72FC04FAF9465883F5E40799274CDE4DB2FC0D4601A868F405E4099BB96900FDA2FC0AB5B3D27BD3F5E40A245B6F3FDD42FC0F6402B3064415E400D8E9257E7D82FC0B1506B9A77405E40679B1BD313D62FC0A5DAA7E3313F5E400B5EF415A4F92FC0971C774A07435E409B3DD00A0C992FC0747B4963B43E5E40990D32C9C8F92FC0352905DD5E425E40EFFE78AF5A992FC0C30DF8FC303E5E406DC5FEB27BF22FC06F2F698CD6415E409B559FABAD982FC082734694F6425E40ECC039234A9B2FC0FF21FDF675405E40EFFE78AF5AD92FC03D0AD7A370415E40672C9ACE4EC62FC081CF0F23843F5E404FAF946588C32FC0295C8FC2F5405E40D21DC4CE14BA2FC024287E8CB93F5E40E89FE06245CD2FC07C7E18213C425E40DEC83CF207C32FC09C8A54185B405E4074EFE192E3CE2FC0C8EF6DFAB33F5E40CE3637A627CC2FC073637AC212435E40EB1C03B2D7BB2FC06EDDCD531D425E40105839B4C8B62FC02BA4FCA4DA3F5E40A8C64B3789C12FC0B03DB32440415E40F146E6913F982FC00D897B2C7D445E40AB048BC3999F2FC0D5264EEE77405E405ABBED4273FD2FC0B2683A3B19445E4075029A081B9E2FC095826E2F694C5E4071C971A774502FC046B6F3FDD4445E404030478FDF9B2FC0143FC6DCB5485E4027F73B14057A2FC0A779C7293A465E40BE30992A18952FC0CE531D7233445E4001DE02098A9F2FC01E5036E50A435E40EBE2361AC0BB2FC0BA490C022B435E406BF12900C6B32FC02176A6D079415E40CE88D2DEE0AB2FC0EB6E9EEA90435E40C3F5285C8FA22FC02506819543435E4086AC6EF59CB42FC0B610E4A084415E407AC7293A92AB2FC0DDEF5014E8435E406B60AB048BA32FC046425BCEA5405E40EA95B20C71AC2FC0EE42739D46425E400D54C6BFCF982FC0EFFE78AF5A415E40EB56CF49EF9B2FC01D554D1075435E40DC2E34D769A42FC0E674594C6C425E40664E97C5C4A62FC08F368E588B435E40F20C1AFA27982FC0A6272CF180425E40B6847CD0B3992FC0B7EEE6A90E455E40107A36AB3E972FC0075F984C15445E404434BA83D8992FC00D897B2C7D445E40E84D452A8C8D2FC0DE9387855A435E400395F1EF338E2FC06E5166834C425E4009168733BF9A2FC0799274CDE4475E40BB270F0BB5662FC0B476DB85E6425E40CBB91457959D2FC0E5B33C0FEE4A5E406C5B94D920532FC072F90FE9B7435E4091ED7C3F359E2FC0397F130A114C5E402254A9D9034D2FC06FF59CF4BE455E40CD920035B56C2FC0F0C4AC1743415E40D42B6519E2982FC0D044D8F0F4425E40922232ACE28D2FC07A53910A63435E404DBED9E6C6942FC03D49BA66F2415E40FC1873D712922FC09487855AD3445E40A1BE654E97852FC0006F8104C5435E4011DF89592F862FC00569C6A2E9445E4027A089B0E1892FC04D2D5BEB8B445E4058CA32C4B18E2FC0DE718A8EE4465E40E63FA4DFBE8E2FC059A31EA2D1455E4067F2CD3637862FC03485CE6BEC465E4089EFC4AC17832FC038842A357B445E409352D0ED258D2FC0ED647094BC465E4075CDE49B6D8E2FC0E09C11A5BD455E404A29E8F692862FC0D0ED258DD1465E4036EA211ADD812FC082AD122C0E475E405EBA490C028B2FC08369183E22465E405648F949B58F2FC0A4703D0AD7435E40AFCE31207B7D2FC09F76F86BB2425E401A69A9BC1D812FC029ED0DBE30455E40CEA5B8AAEC7B2FC0041C42959A415E40609335EA219A2FC08E06F01648485E403524EEB1F4612FC0813E912749435E40FFCA4A9352902FC0102384471B3F5E400BD28C45D3F92FC0FB7953910A435E40EBE2361AC09B2FC0DCF4673F52445E4003780B24289E2FC0492EFF21FD465E40535C55F65D912FC045F5D6C0564D5E40849ECDAACF552FC04B9352D0ED495E40C442AD69DE712FC02A5778978B4C5E4088BA0F406A532FC03E7958A835495E40A301BC0512742FC0BCAE5FB01B4A5E40300DC347C4742FC07E52EDD3F1445E402D431CEBE2762FC0AFCE31207B495E40E6AE25E4835E2FC0B8AF03E78C485E40C1CAA145B6732FC09B030473F4445E400EBE30992A782FC06F641EF983495E40757632384A5E2FC04C1AA37554495E4048BF7D1D38672FC092CB7F48BF455E40CBF8F719176E2FC061545227A0455E40E466B8019F7F2FC082E7DEC325475E400E10CCD1E3772FC04A46CEC29E465E403A58FFE7307F2FC06BD44334BA475E4075C8CD70037E2FC0486DE2E47E475E404C7155D977852FC047C9AB730C445E405305A3923A812FC0A60F5D50DF465E40F12900C633682FC06519E25817575E400E32C9C859182FC03F8C101E6D4C5E4015527E52ED532FC0172B6A300D575E401B0DE02D90202FC094A46B26DF4C5E409E29745E63572FC0DA20938C9C4D5E40D717096D39572FC06B9A779CA24B5E4058CA32C4B16E2FC076374F75C84D5E405ABBED42735D2FC01D386744694B5E40CF4E0647C96B2FC04209336DFF4A5E4008AC1C5A645B2FC031992A18954C5E40B6847CD0B3592FC089981249F44E5E40DC68006F81442FC06C3EAE0D154F5E407407B133854E2FC08811C2A38D4B5E40CD58349D9D6C2FC01E6D1CB1164B5E40EB6E9EEA905B2FC0FF78AF5A994C5E405F419AB1685A2FC034BF9A03044F5E406ABC749318442FC02CF180B2294F5E40917EFB3A704E2FC0F8AA9509BF4C5E40034356B77A4E2FC07EE36BCF2C4D5E40BF0E9C33A2542FC05FD218ADA34A5E405FB532E1975A2FC07689EAAD81495E40E40F069E7B6F2FC03B014D840D575E40840D4FAF94252FC00C1F1153224D5E40CC7F48BF7D5D2FC0FA0AD28C45575E40EACF7EA4882C2FC0F0F96184F04C5E4087F9F202EC632FC02CD49AE61D575E408D7A884677302FC0377172BF43515E40FD304278B4512FC08B89CDC7B5555E406F2A52616C412FC0C91F0C3CF7525E400B462575023A2FC0D00F238447535E40C3F5285C8F422FC04BE5ED08A7515E4082AD122C0E472FC005FA449E24515E40ECC039234A3B2FC09604A8A965535E40C39E76F86B322FC0DBA2CC0699545E40DC63E94317342FC0BBD573D2FB525E4098512CB7B43A2FC01D7233DC80535E40C3BB5CC477422FC07D224F92AE515E40BA490C022B472FC0F0C4AC1743515E407958A835CD3B2FC0C971A774B04E5E405726FC523F4F2FC076711B0DE0515E408FA50F5D503F2FC0DA38622D3E515E404C37894160452FC0D26F5F07CE515E40F168E388B5382FC05B25581CCE505E40F08AE07F2B392FC017D4B7CCE9525E404B598638D6452FC03C8386FE09525E402670EB6E9E4A2FC0CDE49B6D6E545E408048BF7D1D382FC0DC291DACFF535E403ACC9717603F2FC0D482177D05555E40A88C7F9F71412FC073637AC212475E40BF4868CBB9742FC0FB96395D16435E4022718FA50F9D2FC03A234A7B83475E402EFF21FDF6752FC04087F9F202445E403D7E6FD39F9D2FC0EB39E97DE3475E40F2D24D6210782FC07C2C7DE882465E40679B1BD313762FC00A68226C78425E40EDB60BCD759A2FC0AE9E93DE37465E401904560E2D722FC0A06CCA15DE415E4046CEC29E76982FC012A0A696AD455E4017BCE82B48732FC00FEECEDA6D4F5E400C93A98251492FC021938C9C85555E4063B9A5D590382FC0ADFA5C6DC5565E40ACE28DCC231F2FC0AEB6627FD9555E40543541D47D202FC061C3D32B65555E400E4A9869FB372FC0200C3CF71E565E40481630815B372FC09E245D33F9565E4039622D3E05202FC012A0A696AD555E40742497FF901E2FC04AEF1B5F7B565E403108AC1C5A242FC08B1A4CC3F0555E409F1F46088F362FC05F984C158C565E40185B087250222FC038BEF6CC92545E40F20703CFBD272FC071C971A774545E408C67D0D03F212FC029E8F692C6545E406FF59CF4BE312FC0156F641EF9535E404850FC1873372FC03255302AA9535E40AD510FD1E82E2FC034A2B437F8525E4049A297512C372FC07CF2B0506B525E40D8F0F44A59662FC0D847A7AE7CDE5D408E06F016480032C09947FE60E0BD5D40A301BC0512A432C0EBFF1CE6CBBB5D407F6ABC7493A832C05AD8D30E7FBD5D405A643BDF4FAD32C0A9A44E4013BD5D40A8C64B3789B132C0C0266BD443BC5D405D6DC5FEB2AB32C0E4DA5031CEBB5D402041F163CCAD32C04D672783A3BC5D404772F90FE9A732C038A1100187B05D40176536C824B332C04C1AA37554BD5D40D3DEE00B93A932C02A5778978BB05D4068226C787AB532C0D3872EA86FBD5D404165FCFB8CAB32C0068195438BB05D40F12E17F19DB832C06002B7EEE6BD5D40E813799274AD32C0543541D47DBC5D40BD18CA8976B532C04B02D4D4B2BD5D40D656EC2FBBB732C0857CD0B359BD5D4024B4E55C8ABB32C0742497FF90BE5D40F5DBD78173B632C0CBF8F71917BE5D40C898BB9690BF32C00A80F10C1ABE5D407C2C7DE882BA32C0744694F606BF5D40931804560EBD32C0B2F4A10BEABB5D40E6797077D6BE32C06F2F698CD6BD5D40B75D68AED3B832C0494BE5ED08BF5D4006BB61DBA2BC32C03CF71E2E39BE5D40E5F21FD26FBF32C0EDF5EE8FF7BA5D4039622D3E05C032C0C74B378941BC5D405F5E807D74BA32C0DD0720B589BB5D4005C58F3177BD32C01344DD0720BD5D405DFE43FAEDBB32C05322895E46BD5D40C7D79E5912C032C06AF6402B30C05D40AE47E17A14BE32C0EB8B84B69CB35D400B24287E8CC932C07DCB9C2E8BB95D40A06CCA15DEC532C08672A25D85B45D40B459F5B9DACA32C026C79DD2C1B25D4044696FF085C932C0EBAD81AD12C05D40E96514CB2DBD32C02159C0046EB15D40478FDFDBF4C732C0481B47ACC5BF5D40B341261939BB32C020D26F5F07BA5D409014916115BF32C0,VMQ9273,Solander
