+++ pipeline/talend_test/test_configs/soop_trv/soop_trv.yaml
@@ -1,0 +1,25 @@
+
+
+name: soop_trv
+type: pipeline_version_2
+po: laurent
+actions:
+  - type: ADD
+    files:
+    - dest: /mnt/ebs/incoming/SOOP/TRV
+      local_file: IMOS_SOOP-TRV_S_20190328T191556Z_VMQ9273_FV01_END-20190407T215725Z.nc
+drop_schema_objects: true
+run_talend_liqui: true
+talend_log_file: /mnt/ebs/log/pipeline/process/tasks.SOOP_TRV.log
+talend_jobs:
+  - soop_trv-soop_trv
+database_schemas:
+  -
+    name: soop_trv
+    tables:
+      - name: soop_trv_trajectory_data
+        exclude_columns: []
+      - name: soop_trv_trajectory_map
+        exclude_columns: [colour]
+
+
